request #24689: Enable TypeScript incremental builds
On my environment it saves about 1mn in user time when building the core
and all the plugins. The gain are less interesting when looking at real
time: only a few seconds are saved. This is likely due to the fact that
some large part of the build are not optimized by this change and it is
heavily parallelized. Since it has a negligeable impact on clean build
performance it is still an interesting improvement especially when
building a smaller part of the codebase.

Change-Id: I0a065e0047b67cb1915d63a1490b9ffb7ee47e70