chore: Introduce the function to build the Step Definition field value
This is part of story #23462: export test plan as a docx document

No functional change expected here.

TestPlan will need to enhance the StepDefinition field content to
add the executions results. This technical refactoring add a function
that will be called to do that.
For now, the function is the same on both apps.

Change-Id: If3162ece4029a85e4a0e7352f351ed21b40b4de3