Display the icons associated with each tool in the project sidebar
The font-face rules cannot be loaded only inside the Shadow DOM at this
time [0] so we are forced to include them globally. We might encounter
conflicts if another FontAwesome is loaded so we use a custom
font-family name to avoid that.

You can test the change by going into the folder and running `npm run
dev`.

Part of request #23465 Publish a project sidebar custom element

[0] https://drafts.csswg.org/css-scoping/#shadow-names

Change-Id: I806cbcad7b0f4464011a4f52a706681c56b7984a