request #24236 On expert query, search can fail if the searched user doesn't submitted artifact
On expert query, search can fail if the searched user doesn't submitted artifact, even if the user is registered to Tuleap or project member.

The query return "The value 'tlp1234771' doesn't exist for the list field 'submitted_by'." when it should work without result. If there is other criterion, it breaks the whole search.

Change-Id: I75c428b656619fbc2c5db376dcb9458d17d49122