Fix the build of project sidebar component on clean sources
@tuleap/tlp-popovers is now rebuilt inside the sidebar component.
Using it as a dev dependencies does not mark it as a required dependency
when building all the packages so it might not (order might change a bit
due to the concurrency) be built when we start building the project sidebar.

This is not optimal but likely good enough for current needs. This might
be revisited when rebuilding those dependencies becomes too costful.

Part of request #23465 Publish a project sidebar custom element

Change-Id: If2be317e02d29d34a53d79f87da9c0507902a751