feat: Display test result status
This is part of story #23462: export test plan as a docx document

The execution result is now displayed after all the steps.
The execution comment will be added in a dedicated commit.

This commit also adds a fix: the last execution was only retrived for
the traceability matrix, this is now also filtered to display the test
steps status and test result.

Change-Id: I40eb80bfe8b2e9d6ac06a0e057be840a1b30e157