feat: Tests execution attachments in Tests section
If attachments have been added while executing a tests, then they are
listed in a new line "Attachments" in the test results. The link
allows the user to download the corresponding file.

No functional change in test report document generation.

Part of story #23462: export test plan as a docx document

Change-Id: I6d268337241d71d9c580a9b96a333bf5b817836e