feat: Test execution comment in Tests section
The result comment associated to a test execution is now displayed in
the Tests section. If the comment contains inline images (via DnD or
copy/paste in the editor) then they are part of the document as well.

Since we limit ourselves in putting HTML in table cells, this commit
break the table in order to have a consistent reading of the results.

1-cell table for status is kept because it is as of today the only way
to highlight text with random hexadecimal color (highlight property
seems to only accept color names like "red", "blue", …).

1-cell table for attachments is kept because I didn't find a way to have
a paragraph with a bunch of ExternalLink yet. It will be done later.

Part of story #23462: export test plan as a docx document

Change-Id: Iad2a5e5998a4d7b6de5aeab7028f2f6037ed9a46