request #24239 At the artifact creation, data shouldn't search the last value
At the artifact creation, if no value is given for a field date, it will search the last value with wrong data in parameters. It can cause an exception if global rules are set for 2 date, and it found something that have no relation with the actual artifact.

To reproduce :

    Create a tracker with 2 dates and Global rules date_1 < date_2
    Make those field as unsubmitable
    Create an entry in the table tracker_changeset_value_date for each field : as changeset_value_id, set the id of each field.
    Create an artifact

--> you should have the following error :
PHP message: PHP Warning:  preg_match() expects parameter 2 to be string, array given in /usr/share/tuleap/plugins/tracker/include/Tracker/Rule/Date.class.php on line 117PHP

Change-Id: I391513e2faa48b934b5fa1f0439b61a46ef861d8