Do not rebuild @tuleap/tlp-popovers in @tuleap/project-sidebar
This was done this way because our current topological build does not
build dev deps first (see 2c95c7490263d68853c8c399b4cc2014e2503ead).
It is however not a great thing and it might lead to issues in the
future because build settings might be different between the two
packages (e.g. different TS options, specific build tool plugin...).

This contribution introduces Turborepo [0] to manage the build of the
whole repo. It is properly capable to build the @tuleap/* packages in
the expected order. It will also give us the possibility to cache our
build results to speed up rebuilds in dev environments, this is not yet
enabled because it requires to rework a bit how we output our code.
Nx [1] was also considered for the same task but it is less easy to
introduce progressively and has a steeper learning curve.

Part of request #23465 Publish a project sidebar custom element

[0] https://turborepo.org/
[1] https://nx.dev/

Change-Id: Ie8cbf043d82cd9297d628229c52b97c52fea2938