Cache test build outputs
At this stage this is mainly a demo to show how the whole thing works.

To test run `pnpm run test` at the root of the repository, the first
execution will be slow. Once it is done run again `pnpm run test`:
cached test results will be used 🚀. You can do a small change in one of
the app (e.g. add a comment), if you run `pnpm run test` only the needed
part of the tree will be rebuilt.

This change has no impact in the CI pipeline since the cache is trashed
between runs.

Also note that the dependency between the 'test' stage and the 'build'
stage has not indicated in the Turborepo pipeline since the 'build'
stage cannot be cached for now.

Part of request #24241: Cache the build results of the JS toolchain

Change-Id: Ib0863cc2847db8dcb11c3515abef70541a867fa6