Fix loading the system fonts
Until now this was done with directives specific to macOS.
We broke that in 6d924b51e8d283f91c44743e4ae899e66a67db01 because it
appears to care about the case sensitivity.

In any case using vendor directives is not fine so this contribution
fixes it by using the standardized 'system-ui' font-family.

https://drafts.csswg.org/css-fonts-4/#system-ui-def

Part of request #24234 stylelint: 13.12.0 -> 14.2.0

Change-Id: Ic119fda8ac078768d0fccd5572778d647b1587d3