chore(tracker-xml&jira): XMLFormElement no longer depends on IDGenerator
The dependency on IDGenerator implies a verbose API and makes generated
XML unncessary complex (hard to read & debug). This first step is about
having the possibility to let Tuleap generate the IDs for the fields.

The design choice was to rely on Late State Binding (usage of static::)
so the implementation can be mutualized in top level classes. It is not
possible to have a single implementation in XMLFormElement ATM because
we still need XMLFormElementImpl for backward compatibility. It's a
small duplication of the logic but the resulting code is more
straightward.

Finally, there is a need to add dependency on XML tracker in the name
generation to ensure uniquness of IDs when several trackers are being
generated (import project from jira for instance).

IDs are still needed for List fields static values, it will be handeled
separately.

Part of: request #24247 XML generation intermediate representation should not depend on IDGenerator

Change-Id: I1de220467719c3c1d916c93c66ced50e3942e9f0