Drop /account/update-sidebar-preference.php in favor of the REST API
The old endpoint under /account/ cannot be reached outside of the Tuleap
domain (due to CORS policy) so this might become an issue for our
independant sidebar component. In order to keep things similar between
the implementations we start by updating the existing ones to use the
REST API.

Also this old endpoint do not have any form of CSRF protection so there
is really no good reason to keep it.

Part of request #23465 Publish a project sidebar custom element

Change-Id: I89edd535c8c4b4c06680b8fd75e33e94636d0a74