chore(jira): Jira static list values are human readable
The main goal is to have human readable XML output to ease debug as well
as make fixtures more resistant to changes (no XML shuffle because one
new field moved all ids).

The import process (XMLImportFieldStrategyList) had to be changed
because despite the RNC allows value IDs to be string, the code was
enforcing int.

For Jira Import part, the Status values remains IDs because there is too
much coupling with the way status values are exported for semantic and
reports and that would require another mega-refacto.

Closes request #24247 XML generation intermediate representation should not depend on IDGenerator

Change-Id: I85a7bd7dd34fcdfad0b6cf0a65fd7d2ec882a8a4