Sidebar component can be collapsed
In order to give consumers of the component control on how the sidebar
behaves the information is given to the component with the 'collapsed'
attribute.
To update the attribute (as recommended by the best practices when writing
a custom element) when the user clicks on the button we have been forced to
rewrap Vue API.

Since it is not possible for the component to know how to interact with
the Tuleap API, updating the user preference will be under the
responsability of the consumer.

Part of request #23465: Publish a project sidebar custom element

Change-Id: Ibaaf6878d1614b010f9185fe58b1f4a7ef8fc647