Shortcuts can be used with the sidebar component
Hopefully the shortcuts are uniq enough to not conflicts with shortcuts
of the app where the sidebar is used.

You can test the change by going into the folder and running `pnpm run
dev`.

For the sake of consistency between all the sidebars, a small bug have
been fixed in the FP and BP sidebars. When using the `shift+g` shortcut
to select the first tool of the sidebar it is always the first tool that
is focused and not the first tool provided by a plugin.

Part of request #23465 Publish a project sidebar custom element

Change-Id: I6362e856f93a139c1d43dbd4e92f4327af56dcc0