feat(jira): Add support of versisons and fixVersion types
Another monster field as we like them. In apperance everything is fine
fixVersions and versions, that's multiselectbox fields with values &
all.

But the tricky beast is in the management of history. It's not at all
handeled like a regular multiselect value. Whenever you add a new
version it's {from: null, to: "10013"}, the from is null even if there
where versions already defined. On the contrary, removing a version is
mapped with {from: "10003", to: null} and, again, when you encounter
that it doesn't mean that last version was removed but only that version
10003 was removed...

As this completly change the way history is rebuilt, I decided that at
this stage it was not worth it.

This patch comes with a new fixture that highlight the versions
manipulations.

Closes request #24223 Jira import should take versions into account

Change-Id: Ib07c5a231f001fe74b2d05aaff3ac0cb3a20aca3