feat: project text file export is deprecated
The feature is planned to be removed in June 2022. We should warn end
user about this deletion.

I took the opportunity to convert this front page to BurningParrot. This
will help to add project xml export later. Only basic
styling is used in the new table since it will be deleted.

Part of request #24983: Removal of legacy project data export

Change-Id: I5b611e72d3492b609bff03dba8379d1fb1d8d8f7