feat: Project XML export from web UI
Project admin can export the project structure as an XML archive
directly from the project administration (in project data export),
nearly like it is done in CLI[1].

Notes:
* Users and data are not part of the export.
* CLI does not offer yet the option to export only the structure.
* Widgets linked to data (e.g Docman Viewer) are not exported.

Agiledashboard configuration is not exported yet. Therefore the feature
is hidden behind a feature flag. In order to be able to see the export
button:
* Go to project » admin » data » project export
* Add ?feature-flag-xml-structure=1 in the url
* The new button should be displayed

Functional changes to CLI: if --all is not given as option, then
documents and git are not anymore exported.

Part of request #25347: Export project structure from Web UI

[1]: https://docs.tuleap.org/administration-guide/projects-management/export-import/project-export.html

Change-Id: I5b6afc365f940124fbbf6ddb82b431140bdc56ee