request #24990: Tuleap provides its own recent version of Git
The one provided by the SCL is outdated and does not appear to be really
maintained anymore.

We now build Git from nixpkgs and ship static binaries in a RPM
packages.

In case of troubles we can still temporarily rollback the change by
removing /usr/lib/tuleap/git/ folder and redeploy Gitolite config.
Tuleap code will fallback to the rh-git218 version.

To test:
1. Build the tuleap-git-bin package with `nix-build plugins/git/additional-packages/tuleap-git-bin.nix`
2. Install the resulting RPM inside your Tuleap Web container
3. Redeploy Gitolite config with `tuleap-cfg site-deploy:gitolite3-config`

Change-Id: I1b8ebde4d60911eb66968900a880a87ad1cb0b83