feat: Display bug alongside requirement in TTM
When a test execution is linked to one bug, then this bug is displayed
as a badge like when there is a requirement. When user click on the
badge, it is redirected to the bug artifact.

When the execution is linked to more than one bug, then the badge is not
displayed, because a very next contribution will handle this case by
displaying a common badge "Y bugs linked" instead of having Y badges.

Starts request #24972: Tests can cover multiple requirements

Change-Id: I4bb0c545523e4dbd8e7e8c0234484f115dd97e45