Explicit PHP unit tests coverage under the src/ folder
The different symlinks made by the @tuleap/project-sidebar package take a
long time to be processed. It appears that PHPUnit is not capable to
manage loops when it needs to exclude files [0]. To workaround the issue
we remove the global exclusion of the *DAO.php and *Dao.php files that
triggers the loops.

Running PHP unit tests with coverage should not 12+ hours anymore.

Part of request #23465 Publish a project sidebar custom element

[0] https://github.com/sebastianbergmann/php-code-coverage/issues/849

Change-Id: I5cb9a22f26e35203d9a7f95c5775a1a98915a5d5