fix(jira): Import should ignore missing link types
When import is done with jira_import plugin the needed types are created
before the import.

When import is done through Tracker creation, link types might not be
there.

In both case, when link type is missing it should not lead to an empty
changeset because all the other data are lost.

Closes request #24998 Jira issue type import shouldn't fail on missing link types

Change-Id: Iccf53d9d13b8ef70dc6a72879adc7b7379b4bcd3