Drop indirect dependency to sprintf-js in the project sidebar component
This dependency is required by the Tuleap core for the description of
the shortcut. This is not really something we need in this context and
sprintf-js is not really needed here anyway. Since we, voluntarily,
bundle everything into the component we ship, not having indrect
dependency will ease our work e.g. it is easier to determine if we need
to ship a new version because a vulnerability has been detected in one
of the dependency.

It also save a few KB which is a good thing.

Part of request #23465 Publish a project sidebar custom element

Change-Id: I61f392dea9b62250bce7a6cd393bb583055dc6a2