fix(jira): manage priority when it's not part of exposed fields
On older jira versions (8.5.1) REST API might have a priority in the
data but not present in createmeta nor in editmeta (🙄) so we have to be
clever.

The choice here is to rely on the issue information part of 'editmeta'
payload. As editmeta also have one issue, if there is a priority in the
payload but no priority field already seen we recreate it.

Their is no way to test the behaviour Live (the SBX integrated test
shows what that does with an old version). For newer version (esp. Cloud
nothing should change).

Closes request #24224 Jira import should take priority into account

Change-Id: Ic5b15e66939e68731c59251a90f210595be95a8f