Introduce @tuleap/project-sidebar-internal
This will let us use the project sidebar element inside the Tuleap codebase
without duplicating the heavy fonts.

When accounting for all the elements (code, CSS, HTML code), it is in
fact almost worth it in term of size to re-use the element and it
ensures we provide the same experience everywhere. There will be a small
runtime cost due to Vue but it is negligible.

The explicit dependency to @tuleap/core has been replaced by a relative
path. Accessing the code via @tuleap/core creates a symlink to an upper
directory which can put tools iterating the files of the codebase into a
loop.

No functionnal changes, you can go to both packages and run `pnpm run
dev` to play with them.

Part of request #23465 Publish a project sidebar custom element

Change-Id: If9b47d10f451a33c2f9203c04b179ac4bc926845