Git_ExecTest::testGetAllBranchesSortedByCreationDate relies on an undefined behavior
In recent Git versions when two branches have the same 'creator date'
they are now alphabetically sorted instead of letting an external tool
determine it (in this case I'm guessing the filesystem).

Anyway the test was not testing the correct feature we wanted to test,
this is now fixed.

Part of request #24990: Tuleap provides its own recent version of Git

Change-Id: I7ea9926a86c677b67d81c61ac97d9b191f775b5f