feat: Package gitolite3 for Tuleap
Regular gitolite3 package is fine but depends on `git` package provided
by RHEL/CentOS. Not only it's old (1.8.X on v7) but its usless as it's
been a while since we no longer use it. It might even be confusing for
administrators to have this old git by default.

As for tuleap-git-bin, the version of the package is aligned with the
corresponding Tuleap version so we are sure to deliver a coherent set of
package and not miss an update.

While the package is being built in our regular build chain we took the
opportunity to remove the packaging elements that are no longer needed
in Tuleap context:
* Providing Gitolite3 perl modules (there are no dependencies on them)
* The gitolite3 user and its home dir (neither used)
* Fedora extra files (compile-1 that we don't use and specific Readme
  that nobody read)

In order to work as intended the git setup deployment was updated:
- Until now, setup was still relying on /usr/bin/git to work. To fix
  that site-deploy:gitolite3 must be run twice. One before the setup so
  gitolite environment (via .profile) knows were to find git. And one
  after setup so .gitolite.rc contains the ENV informations for git.
- It seems that (recent?) changes on centos/rhel users generates a
  default ~/.bash_profile file on every account. When present ~/.profile
  is not loaded at all. So I decided to remove ~/.bash_profile is
  present because for gitolite user there is no reasons it should load
  sth usefull from ~/.bash_profile.

If things go wrong with tuleap-gitolite3 package, the revert operation:
- rpm -e --nodeps tuleap-gitolite3
- yum install --exclude=tuleap-gitolite3 gitolite3

Part of: request #25004 Tuleap should ship its own gitolite package

Change-Id: I46c2432959fdfb46fe77c99d96ae92f8ce9aad47