Split checking of Plannable trackers
part of request #24994 : Javascript error in Iterations planning

No functional change expected.

Notes:
Program Management domain is not supposed to know what a Tracker is.
Therefore, we split the code responsible for checking Program Increment,
Iteration and Plannable trackers are valid. This way, the underlying
implementation could change independently. It is in fact still the same
as all three interfaces are implemented by a single adapter that runs
the same code.

Change-Id: I98420b4a792d4e13239b178cd8e2a15b7a473cfc