Remove remaining usages of /usr/bin/git
We do not have a direct dependency to it so we should expect to have it
available, we should always use the Git provided by Tuleap.

A direct dependency to openssh has also been added because we use
`ssh-keygen` and at the moment it is only provided by an indirect
dependency of the tuleap-plugin-git package.

Part of request #24990: Tuleap provides its own recent version of Git

Change-Id: I7aa58dc6e217351cf0182f6283f8d580f5c3ac0d