Split checking of new Program Increment tracker
part of request #24994 : Javascript error in Iterations planning

No functional change expected.

Notes:
The goal is to remove TrackerIsValidChecker static method and to split
checking the validity of Program Increment, Iteration and Plannable
Trackers.

Program Management domain is not supposed to know what a Tracker is.
Therefore, we must split the code that checks trackers are valid. This
way, the underlying implementation could change independently. It is in
fact still checked the same way as all three interfaces will be
implemented by a single adapter that runs the same code.

Renamed objects with "New" prefix and added comments to underline the
fact that those objects exist to be saved in storage.
Added a "SavePlan" interface to remove a mock.

Change-Id: Ib2a3cda58111ea6d76bf47139976cb6c504cf5df