request #25306: Javascript error after closing switch-to modal
How to test:
- go to Tuleap homepage. Hit "c" and then "s" shortcuts. It should open
first the "+ New" dropdown and then the "Switch to ..." modal
- There should not be a javascript error
- "user navigation" end-to-end test should succeed in Tuleap Core.

Notes:
The failing code made the assumption that an element with CSS class
"tlp-dropdown-menu-item" was always inside a TLP dropdown, but it is no
longer true. Navbar dropdowns are not TLP dropdowns but now they contain
elements with that CSS class. The code now checks if it can find a
dropdown_menu, and if not it assumes the active element is not a
dropdown.

Change-Id: I4b2aedc6ea855f6dba4b42536bfaa060fffdb90f