feat(forge-config): Seed docker image with DB passwords
This commit bring the ability to define the Site Administrator password
as an environment variable at container init. In most basic scenario it
means that you can have a container without autogenerated DB passwords.
There are still areas not covered (nss password for instance) that will
be done in later commit. It's not yet possible to change the password of
site admin either.

Most of the work here was done to keep classic setup and container setup
as close as possible.

As stated in the associate Story, we won't be able to only run with
Environment Variables (without variable files like local.inc or
database.inc written down) because it requires too much change (fpm
cleans env, sudo will do it as well, etc) that are not worth the risk of
letting stuff of the radar. So the strategy will be to store in DB
(later) or write down on filesystem (now + "CannotBeModified"
variables).

Finally, I choose to keep the database.inc format even if generated out
of the Environment Variables to limit the amount of changes on
ForgeConfig.

Part-of: story #22708 flexibility in the management of Tuleap Configuration

Change-Id: Id4e8df7eb69a8b770acf507ab7c54a709dc2fc63