feat: Export traceability matrix of campaign in docx
Backlog and tests section are kept empty for now to not clutter to much
the review. It is already quite big: nearly all docx generation from
testplan has been moved to testmanagement since the document content is
the same. The only difference is the content of the cover page to
display the current milestone (and its parent) for testplan and the
current campaign for testmanagement.

We are forced to use our own gettext in document generation since
gettext provider in testmanagement (AngularJS) and testplan (VueJS) does
not share the same interface.

No functional change in testplan document generation is expected.

Part of story #23494: export test campaign as docx

Change-Id: Ia4427dc01437b0d37e43ca3e473ee6cd1f979ede