feat: Export tests of campaign in docx
Tests of the campaign are now part of the document.

Code dealing with this retrieval have been copied from testplan as is
(See `createExportReport`[1]).
We will see if we can factorize something between both apps later, after
backlog of campaign is done.

No functional change in testplan document generation is expected.

Part of story #23494: export test campaign as docx

[1]: plugins/testmanagement/scripts/testmanagement/src/helpers/ExportAsDocument/Reporter/report-creator.ts

Change-Id: I1953de9bb786c5bbd28b0df9d6c2c3a5aff7db7e