Move to CommonJS modules
Using AMD modules is going to be annoying to bundle everything in a
single binary.

ES Modules were also a possibility but it required (to do it cleanly)
to also drop the require() in the tests and it does not bring more
advantages for goal of having a realtime server again.

Part of request #25312: tuleap-realtime package cannot be built or installed

Change-Id: I6ab47c28af3d7d522c1b43afc90e5f9f488316c0