feat(new-art-link-field): Display links table
part of: story #24968 [modal] have the list of linked artifacts

Note: For the moment, links with closed statuses will have a green badge
and are not grayed. This is expected as we do not have the information
if the artifact is open or closed yet. It will be done in a next commit.

How to test:
- In creation mode
--> Still no table

- In edition mode
--> Skeleton when artifacts are being loaded
--> The table is displayed when there are links. It is not displayed
   otherwise (IMO an empty state seems unrelevant here, but we'll ask
   UXD later)

Change-Id: Ia77fb9f7f63a3acd750b99d08f3f4c7809a5b61d