Bundle the realtime server with Node 16
We do not rely on the NodeJS coming from EPEL anymore.
We do the bundling using vercel/pkg [0]. The bytecode generation is
disabled because it is not reproducible and it also a PITA to do
without building Node from scratch in a pure Nix environment.

You can generate the executable with `pnpm run build` and you can build
the RPM with `OUTPUTDIR=/tmp OS=centos:7 ./packaging/build-rpm.sh`.

Part of request #25312: tuleap-realtime package cannot be built or installed

[0] https://github.com/vercel/pkg

Change-Id: I50402530ccc836dc803a1c71c06bbdfd19947194