Disable binary stripping automatically done by rpmbuild
Trying to strip the binary will remove the comment section where the
virtual filesystem containing the code of the app is stored. This
prevent the executable to be launched since it cannot find the files to
execute anymore.

Trying to execute the binary stored in the RPM package should not throw
`Pkg: Error reading from file.`.

Part of request #25312: tuleap-realtime package cannot be built or installed

Change-Id: Ia02890e5bdc1bc674d327f49254cf7433acfafbb