Show "parent feedback" with Hexagonal Architecture
part of: story #24968 [modal] have the list of linked artifacts

How to test:
1. set the config flag "feature_flag_use_art_links_v2_in_artifact_modal"
to 1
  - Rebuild the agile dashboard and artifact modal.
- Go to planning view and open the modal to add a child to a user
story.
- You should see a feedback at the top of the modal like "The artifact
will be linked to <Parent title>"

2. set the config flag to 0
  - Open the same modal.
- You should see the same feedback. The parent artifact is expected to
be requested twice.

Notes:
Avoiding the double request without the feature flag means tinkering a
lot with the "legacy" link field controller. The aim of the story is to
kill it, so we consider it acceptable to temporarily request twice and
later remove the "legacy" link field completely.

We can use Hexagonal Architecture in the frontend too. We don't even
need classes: closures, plain objects and interfaces do the job !

Change-Id: Ic9942b27038600afe7dca80b6198c1a519d843a4