feat: Display search criteria panel
Note: You have to activate the feature beforehand:
tuleap config-set feature_flag_search_for_documents_with_criteria 1

Go to Document interface and in the search box type some terms and hit
enter. You reach a search result page with a search criteria panel (with
search terms displayed in a "Global search" field), and an empty table
(results will come later in a dedicated contribution).
This page is bookmarkable (search terms are part of the URL).

In the previous search box, if you click on the `Advanced` link, then
the search terms are empty, and the results table is not displayed.

In the search criteria, if you enter another search terms and apply the
search, then you end up in a new result page (still bookmarkable).

Part of story #24197: search for Documents with criteria

Change-Id: I910933691dd422947d94f223b7c64a85a88e244c