Make possible to use <script setup> in Vue 2 component
Note that I had to keep the basic <script> to avoid issues with
ForkTsCheckerWebpackPlugin. This could be avoided with vue-tsc but
switching of typechecking is not the goal of thi set of patches.

Also, unplugin-vue2-script-setup/jest was not used to configure Jest
because it still uses vue-jest (I will contribute a patch upstream).

Part of request #25317 Demonstrate how to use Composition API in Vue 2 app

Change-Id: I9cf8b438819999a9013d9d817ab641a004f2083e