feat: Start to display search results
If some items are matching the query, they are displayed in the table.

Limitations that will be managed in dedicated contributions:
- No icon
- Description column is not interpreted
- No location computation
- No pagination

Note: You have to activate the feature beforehand:
tuleap config-set feature_flag_search_for_documents_with_criteria 1

This includes the bump of prettier in order to [be able to create
snapshots][1]. See [prettier 2.5.1 changelog][2] for details.

Part of story #24197: search for Documents with criteria

[1]: https://github.com/facebook/jest/issues/12112
[2]: https://github.com/prettier/prettier/blob/2.5.1/CHANGELOG.md

Change-Id: Ibc094812bf0d37350695c91cb6feb6200c875559