chore(forge-config): Generate database.inc without .inc.dist
In order to be able to generate the database.inc file without doing
query/replace in .inc.dist file we need to be able to describe the
variables completly using attributes. Hence introduce of default values
and config help attributes.

For default values, I propose to associate them to the a new attribute
that define ConfigKey type. In the various attempts around this idea it
seems to me that it's a good compromise between the amount of code to
write for each config key and the ability to have working code when
dealing with the config variables. An alternative can be seen
here: Idadb114ddf06fb0c547211f524ed4089ad81a020

There is a special default value for "LegacyBool" because I do not want
to break existing code that might rely on bool being strings ('1').

Part-of: story #22708 flexibility in the management of Tuleap Configuration

Change-Id: If6a17aee930debe7e2ed93b3fc25f7f35b6be891