feat: Paginate search results
When your search result in documents return results, only 50 fist items
are displayed and pagination buttons are displayed allowing you to
browse every results.

If user starts to search something, then pagination is reset to 0 (even
if it is the same terms).

The URL contains the offset, so that users can send the result page to
their BFF.

Note: You have to activate the feature beforehand:
tuleap config-set feature_flag_search_for_documents_with_criteria 1

Part of story #24197: search for Documents with criteria

Change-Id: Idd8c780d9e54afb2491c571c6bfc6b786dad9929