feat: remove docman administration frontpage
Users coming from Document interface should not gag when they reach the
docman administration. The front page is therefore removed in favor of
tabs, like it has been done elsewhere in Tuleap.

The first tab is the default one. As of today it is permissions.

This is a first step to be able to introduce Search administration page.

Note: the link to "Search report" is removed from administrations
context since this page is accessible from reports by everyone. We
didn't find a way to include it in administration pages **and**
available for everyone. So let's remove it and see how they react in the
field.

Part of story #24197: search for Documents with criteria

Change-Id: Ib5111db2c0851169086471f76ed007d89e306b5e