request #25328: Anonymous Git over HTTP operations do not succeed
The issue is caused by the fact that
/usr/lib/tuleap/git/libexec/git-core/git-http-backend try to find
Git in the Nix store and when it do not, it fallbacks to whatever Git
version is available in the path.

In order to avoid that we change a bit the way we build our static
binaries. Instead of rewrapping the Git binary to set some env variables
to set the appropriate paths we tell the Git build system to use our
final deployment path as a prefix and to "install" inside the Nix store.
This is different than the base derivation where the "$out" path of the
Nix store is used as the prefix. The resulting binaries do not expect to
find the other parts inside the Nix store but under /usr/lib/tuleap/git
as we expect.

Change-Id: I90e83526b52f673110b3fea3ea34bc9d77a7bb32