Move Controller and Presenter to Adapters
part of: story #24968 [modal] have the list of linked artifacts

No functional change in the modal feedback showing the future parent of
the artifact.

Notes:
Presenter is highly dependent on UI and should therefore be in Adapters.
Controller is tied to a UI Component and builds Presenters, so it will
also need to be in Adapters.

Change-Id: Ib81389e57f854d7f5285b7fed23bf59b6e6a995d