fix: Enhance handling of standard fields in legacy tracker migration
This fixes request #25330 Enhance handling of standard fields in legacy tracker migration

When standard fields have values stored in the wrong SQL table (this happen in one case),
then this value is falsely taken into account.

How to reproduce the issue
--------------------------

* Add a default value to a standard field like summary
* Disable this field
* Enable this field
* Do a TV3 to TV5 migration

All the migrated artifacts (that had some changes in TV3) will have the default value in the TV5,
which is wrong.

Change-Id: Ib4824db8e5edd2d9a51e34bbedd7a4de743bdac3